Split mobile and desktop versions of loading screen
